LayoutTests/fast/events/wheel/platform-wheelevent-paging-xy-in-scrolling-page-expected.txt

PASS event.wheelDeltaY is window.givenScrollTop * -3
PASS event.wheelDeltaX is window.givenScrollLeft * -3
PASS event.wheelDelta is window.givenScrollTop * -3
PASS document.scrollingElement.scrollTop is window.expectedScrollTop
PASS document.scrollingElement.scrollLeft is window.expectedScrollLeft
